Move std::num::Integer to libnum
